requirements.txt
langchain
langchain-community
chromadb
sentence-transformers 
transformers
openai
tiktoken
pandas
lark
streamlit
python-dotenv
langchain-pinecone
pinecone